feature  : 제작 및 건축 시 아이템 사용 / ui 추가

+++ Source/ProjectOpenWorld/Private/Inventory/Widget/InventoryWidget.cpp
@@ -93,8 +93,8 @@ void UInventoryWidget::UpdateAllInventorySlot()
 
 void UInventoryWidget::UpdateInventoryWeight()
 {
-	int MaxWeight = inventoryComponent ? inventoryComponent->GetInventoryMaxWeight() : 0;
-	int CurWeight = inventoryComponent ? inventoryComponent->GetInventoryWeight() : 1;
+	float MaxWeight = inventoryComponent ? inventoryComponent->GetInventoryMaxWeight() : 0;
+	float CurWeight = inventoryComponent ? inventoryComponent->GetInventoryWeight() : 1;
 	WeightProgressBar->SetPercent(CurWeight / (float)MaxWeight);
 	WeightPercentTextBlock->SetText(FText::Format(FText::FromString(TEXT("{0} / {1}")), CurWeight, MaxWeight));
 }
